ProtonSettings: Move proton search function into `utils/proton.py`
